+++ src/main/kotlin/com/oksusu/susu/event/model/BaseEvent.kt
@@ -1,4 +1,4 @@
-package com.oksusu.susu.common.dto
+package com.oksusu.susu.event.model
 
 import java.time.LocalDateTime
